pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundryiq-demo"
version = "1.0.0"
description = "AI FoundryIQ Demo - Document Q&A with Azure AI"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "azure-search-documents>=11.4.0",
    "azure-identity>=1.15.0",
    "openai>=1.12.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.0.0",
    "PyPDF2>=3.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
